app/requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
python-socketio==5.10.0
eventlet==0.33.3
gunicorn==21.2.0
python-dotenv==1.0.0
stripe==7.8.0
phonenumbers==8.13.27
geopy==2.4.1
requests==2.31.0
Pillow==10.1.0
pytz==2023.3
cryptography==41.0.8
bcrypt==4.1.2
email-validator==2.1.0
qrcode==7.4.2